# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A link shortening service with caching in Redis",
        "title": "URL Shortener API",
        "contact": {},
        "version": "1.0"
    },
    "host": "Localhost:8080",
    "basePath": "/",
    "paths": {
        "/resolve/{code}": {
            "get": {
                "description": "It accepts a short code and redirects to the original URL",
                "tags": [
                    "urls"
                ],
                "summary": "redirect using a short code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Empty code",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Code not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shorten": {
            "post": {
                "description": "It takes the original URL and returns a short code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "create a short link",
                "parameters": [
                    {
                        "description": "Original URL",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ShortenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Short code",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or empty URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ShortenRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        }
    }
}